api/pyproject.toml
[tool.poetry]
name = "upoutodo"
version = "0.1.0"
description = ""
authors = ["Marco Enrico <mvalviar@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
cryptography = "^42.0.8"
drf-jwt = "^1.19.2"
requests = "^2.32.3"
pyjwt = "^2.8.0"
auth0-python = "^4.7.1"
django-filter = "^24.3"
django-taggit = "^6.1.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
pre-commit = "^4.1.0"
pytest = "^8.3.5"
pytest-django = "^4.10.0"
faker = "^37.0.2"
factory-boy = "^3.3.3"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "upoutodo.settings"
testpaths = ["upoutodo/api/tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
